+++ src/commands/announce.js
@@ -1,11 +1,14 @@
 import { SlashCommandBuilder } from "@discordjs/builders";
-import { ActionRowBuilder, EmbedBuilder, ButtonStyle, Colors } from "discord.js";
-
-import { createButton, interactionReply } from "./utils.js";
-import { isAdmin } from "../helpers/index.js";
-import { PERSONALITY } from "../personality.js";
-
+import {
+  ActionRowBuilder,
+  EmbedBuilder,
+  ButtonStyle,
+  Colors,
+} from "discord.js";
+import { createButton } from "./utils.js";
+import { interactionReply, isAdmin } from "../helpers/index.js";
 import { COMMONS } from "../commons.js";
+import { PERSONALITY } from "../personality.js";
 
 // GIFT Announce
 const giftAction = async (interaction) => {
@@ -31,7 +34,7 @@ const giftAction = async (interaction) => {
     );
 
   //get channel
-  const server = COMMONS.fetchGuildId(interaction.guildId);
+  const server = COMMONS.fetchFromGuildId(interaction.guildId);
   const channelId = server.announce.giftChannelId;
   const channel = await interaction.client.channels.fetch(channelId);
 
@@ -56,7 +59,7 @@ const action = (interaction) => {
   const announceP = PERSONALITY.getCommands().announce; //get personality
 
   if (!isAdmin(interaction.user.id)) {
-    //check for bot admin
+    //check for admin
     interactionReply(interaction, announceP.notAdmin);
     return;
   }