+++ src/commands/eventRoles.js
@@ -1,14 +1,20 @@
 import { SlashCommandBuilder } from "@discordjs/builders";
-import { ActionRowBuilder, EmbedBuilder, ButtonStyle, Colors } from "discord.js";
+import {
+  ActionRowBuilder,
+  EmbedBuilder,
+  ButtonStyle,
+  Colors,
+} from "discord.js";
 
-import { createButton, interactionReply } from "./utils.js";
+import { createButton } from "./utils.js";
 import {
   addEventRole,
   getEventRoles,
+  interactionReply,
   updateEventRoleMessageId,
 } from "../helpers/index.js";
-import { PERSONALITY } from "../personality.js";
 import { COMMONS } from "../commons.js";
+import { PERSONALITY } from "../personality.js";
 
 export const eventRolesButtonHandler = async (interaction) => {
   const { customId, guildId } = interaction;
@@ -23,7 +29,7 @@ export const eventRolesButtonHandler = async (interaction) => {
   const eventRoleId = currentEventServer[requestedEventRole + "RoleId"];
 
   //get alavirien role id
-  const currentServer = COMMONS.fetchGuildId(guildId);
+  const currentServer = COMMONS.fetchFromGuildId(guildId);
 
   //give requested role
   const guildMember = !interaction.member
@@ -92,7 +98,7 @@ const action = async (interaction) => {
   const options = interaction.options;
   const subcommand = options.getSubcommand();
 
-  const currentServer = COMMONS.fetchGuildId(interaction.guildId);
+  const currentServer = COMMONS.fetchFromGuildId(interaction.guildId);
 
   const db = interaction.client.db;
   const buttonType = ButtonStyle.Primary;
@@ -186,7 +192,7 @@ const action = async (interaction) => {
     const newRoleObj = {
       name: slicedName,
       permisions: baseRole.permisions,
-      reason: `Demandé par ${interaction.member.toString()}`,
+      reason: perso.author + interaction.member.toString(),
     };
     if (color) newRoleObj.color = color;
     const newRole = await roles.create(newRoleObj);
@@ -211,7 +217,7 @@ const action = async (interaction) => {
     //create new button
     const emoteId = name.includes("<") ? name.split(">")[0] : null;
     const newButton = createButton(
-      "eventRole_" + slicedName,
+      perso.prefix + slicedName,
       slicedName,
       buttonType,
       emoteId
@@ -219,15 +225,19 @@ const action = async (interaction) => {
 
     //create new vote buttons + regroup with olders
     const oldComponents = roleMessage.components;
-    const oComponents = oldComponents.reduce((acc, cur) => [...acc, ActionRowBuilder.from(cur)], []);
-    const lastARSize =
-      oComponents[oComponents.length - 1].components.length;
+    const oComponents = oldComponents.reduce(
+      (acc, cur) => [...acc, ActionRowBuilder.from(cur)],
+      []
+    );
+    const lastARSize = oComponents[oComponents.length - 1].components.length;
     const newComponents =
       lastARSize === 5
         ? [...oComponents, new ActionRowBuilder().addComponents(newButton)]
         : [
             ...oComponents.slice(0, -1),
-            ActionRowBuilder.from(oComponents[oComponents.length - 1]).addComponents(newButton),
+            ActionRowBuilder.from(
+              oComponents[oComponents.length - 1]
+            ).addComponents(newButton),
           ];
 
     //edit message
@@ -237,7 +247,7 @@ const action = async (interaction) => {
     });
 
     //reply to interaction
-    if (status && status2) interactionReply(interaction, "c'est bon");
+    if (status && status2) interactionReply(interaction, perso.ok);
     else interactionReply(interaction, perso.errorGeneral);
   }
 };