+++ src/commands/polls/polls.js
@@ -1,13 +1,18 @@
+import dayjs from "dayjs";
 import { SlashCommandBuilder } from "@discordjs/builders";
 import { ActionRowBuilder, EmbedBuilder, ButtonStyle } from "discord.js";
-import { addPoll, addPollChoices } from "../../helpers/index.js";
-import { PERSONALITY } from "../../personality.js";
 import { pollButtonCollector } from "./pollsCollectors.js";
-import { createButton, interactionReply } from "../utils.js";
 import { parsePollFields, stopPoll } from "./pollsUtils.js";
-import { getPollFromTitle, getPollsTitles } from "../../helpers/db/dbPolls.js";
+import { createButton } from "../utils.js";
+import {
+  addPoll,
+  addPollChoices,
+  getPollFromTitle,
+  getPollsTitles,
+  interactionReply,
+} from "../../helpers/index.js";
 import { COMMONS } from "../../commons.js";
-import dayjs from "dayjs";
+import { PERSONALITY } from "../../personality.js";
 
 const command = new SlashCommandBuilder()
   .setName(PERSONALITY.getCommands().polls.name)
@@ -146,7 +151,7 @@ const action = async (interaction) => {
 
   //check for alavirien.ne role
   const guildMember = await interaction.member.fetch();
-  const currentServer = COMMONS.fetchGuildId(interaction.guildId);
+  const currentServer = COMMONS.fetchFromGuildId(interaction.guildId);
   if (!guildMember.roles.cache.has(currentServer.alavirienRoleId)) {
     interactionReply(interaction, personality.errorNotAlavirien);
     return;
@@ -169,16 +174,17 @@ const action = async (interaction) => {
     const voteMax = option == null ? 1 : option;
 
     option = options.getString(perso.colorOption.name, false); //color
-    const color = option == null ? pColors.choices[4].value : option;
+    const color = option == null ? pColors.choices[2].value : option;
 
     const author = options.getUser(perso.authorOption.name, false); //author
 
     option = options.getNumber(perso.hourOption.name, false); //hours
-    const hours = option == null ? 48 : option;
+    const hours = option == null ? 0 : option;
     option = options.getNumber(perso.minuteOption.name, false); //minutes
-    const minutes = option == null ? 59 : option;
-    const timeout = (hours * 60 + minutes) * 60 * 1000; //poll duration in miliseconds
-    const pollDate = dayjs().millisecond(timeout).toISOString();
+    const minutes = option == null ? 0 : option;
+    let timeout = (hours * 60 + minutes) * 60 * 1000; //poll duration in miliseconds
+    if (timeout === 0) timeout = 48 * 60 * 60 * 1000; //2 days default value
+    const pollDate = dayjs().millisecond(timeout);
 
     //check if not too many choices
     const splited = choices.split(";");
@@ -198,10 +204,11 @@ const action = async (interaction) => {
       embed.setAuthor({ name: author.username, iconURL: author.avatarURL() });
 
     //write footer according to voteMax
-    const footerText =
+    const voteFooter =
       voteMax === 1
-        ? perso.footer.unique + perso.footer.options
-        : perso.footer.multiple + ` (${voteMax})` + perso.footer.options;
+        ? perso.footer.unique
+        : perso.footer.multiple + ` (${voteMax})`;
+    const footerText = voteFooter + perso.footer.options;
     embed.setFooter({ text: footerText });
 
     // Optionnal parameters
@@ -212,8 +219,9 @@ const action = async (interaction) => {
 
     //write choices in embed
     const black = personality.black;
-    results.fields.forEach((field) => {
-      embed.addFields({ name: field, value: black.repeat(10) + " 0% (0)\n" });
+    results.fields.forEach((field, idx) => {
+      const name = results.emotes[idx] + " " + field;
+      embed.addFields({ name, value: black.repeat(10) + " 0% (0)\n" });
     });
 
     //create vote buttons
@@ -248,7 +256,7 @@ const action = async (interaction) => {
     );
 
     //add setting button
-    const settingId = "polls_" + "settings";
+    const settingId = personality.prefix + perso.settings;
     const settingButton = createButton(
       settingId,
       null,
@@ -264,10 +272,17 @@ const action = async (interaction) => {
         settingButton
       );
 
+    //add timeout embed
+    const timeoutEmbed = new EmbedBuilder().setColor(color);
+    timeoutEmbed.addFields({
+      name: perso.timeout,
+      value: `<t:${pollDate.unix()}:R>`,
+    });
+
     //send poll
     try {
       const pollMsg = await interaction.channel.send({
-        embeds: [embed],
+        embeds: [embed, timeoutEmbed],
         components: components.actionRows,
       });
       pollButtonCollector(pollMsg, timeout); //start listening to interactions
@@ -285,7 +300,7 @@ const action = async (interaction) => {
         colorIdx,
         voteMax,
         title,
-        pollDate
+        pollDate.toISOString()
       ); //add to db
     } catch (e) {
       console.log("/polls create error\n", e);
@@ -332,16 +347,17 @@ const action = async (interaction) => {
     //add to embed
     const results = parsePollFields(splited, totalSize);
     const black = personality.black;
-    results.fields.forEach((field) => {
-      embed.addFields({ name: field, value: black.repeat(10) + " 0% (0)\n" });
+    results.fields.forEach((field, idx) => {
+      const name = results.emotes[idx] + " " + field;
+      embed.addFields({ name, value: black.repeat(10) + " 0% (0)\n" });
     });
 
     //create new vote buttons + regroup with olders
     const settingButton = lastAR.components[lastAR.components.length - 1]; //get settings button
-    const voteAR = [
-      ...oldComponents.slice(0, -1),
-      lastAR.spliceComponents(-1, 1),
-    ]; //filter actionRows
+    const lastARsliced = ActionRowBuilder.from(lastAR);
+    lastARsliced.components.splice(-1, 1);
+
+    const voteAR = [...oldComponents.slice(0, -1), lastARsliced]; //filter actionRows
     const initComponents = {
       actionRows: voteAR,
       size: voteAR[voteAR.length - 1].components.length,