+++ src/listeners.js
@@ -1,18 +1,154 @@
-import { PERSONALITY } from "./personality.js";
-
-import {
-  reactionHandler,
-  deleteSongFromPlaylist,
-  removeReminder,
-} from "./helpers/index.js";
-
+import { ChannelType } from "discord.js";
 import { presentationHandler } from "./admin/alavirien.js";
 import { roleAdd, roleRemove } from "./admin/role.js";
-
 import { octagonalLog } from "./admin/utils.js";
+import { buttonHandler, selectMenuHandler } from "./commands/utils.js";
+import {
+  addEmojiData,
+  addServerEmojiCount,
+  addServerStatsData,
+  deleteSongFromPlaylist,
+  interactionReply,
+  isReleasedCommand,
+  removeReminder,
+} from "./helpers/index.js";
 import { COMMONS } from "./commons.js";
+import { readContentAndReact } from "./fun.js";
+import { emojiInContentHandler, statsGifCount } from "./stats.js";
+import { PERSONALITY } from "./personality.js";
 
-export const onPublicMessage = (message, client, currentServer) => {
+//#region Listeners
+export const onInteractionCreate = (interaction) => {
+  if (interaction.isButton()) {
+    buttonHandler(interaction);
+    return;
+  }
+
+  if (interaction.isStringSelectMenu()) {
+    console.log("selectMenu interaction detected");
+    selectMenuHandler(interaction);
+    return;
+  }
+
+  const client = interaction.client; //get client
+
+  if (interaction.isContextMenuCommand()) {
+    //context commands
+    const contextCommands = client.contextCommands; //get commands
+
+    const foundCommand = contextCommands.find(
+      (cmd) => cmd.command.name === interaction.commandName
+    );
+
+    if (foundCommand) foundCommand.action(interaction); //if found command, execute its action
+    return;
+  }
+
+  const slashCommands = client.slashCommands;
+
+  if (interaction.isAutocomplete()) {
+    //interaction with autocomplete activated
+    const autoCompleteCommands = slashCommands.filter(
+      (cmd) => cmd.autocomplete
+    ); //get commands with autocomplete action
+    const foundCommand = autoCompleteCommands
+      ? autoCompleteCommands.find(
+          (cmd) => cmd.command.name === interaction.commandName
+        )
+      : null; //find command that fired onInteractionCreate
+    if (foundCommand && isReleasedCommand(foundCommand))
+      foundCommand.autocomplete(interaction);
+    else interaction.respond([]); //if not found, return no choices
+  } else if (interaction.isCommand()) {
+    //slash commands
+    const client = interaction.client; //get client
+    const slashCommands = client.slashCommands; //get commands
+
+    const foundCommand = slashCommands.find(
+      (cmd) => cmd.command.name === interaction.commandName
+    );
+
+    if (foundCommand && isReleasedCommand(foundCommand))
+      foundCommand.action(interaction, "/");
+    //if found command, execute its action
+    else
+      interactionReply(
+        interaction,
+        PERSONALITY.getAdmin().commands.notReleased
+      );
+  }
+};
+
+export const onMessageCreate = async (message) => {
+  // Function triggered for each message sent
+  const { channel } = message;
+
+  if (channel.type === ChannelType.DM) return;
+  else {
+    const currentServer = COMMONS.fetchFromGuildId(channel.guildId);
+    onPublicMessage(message, currentServer);
+  }
+};
+
+export const onReactionAdd = async (messageReaction, user) => {
+  // Function triggered for each reaction added
+  const currentServer = COMMONS.fetchFromGuildId(
+    messageReaction.message.channel.guild.id
+  );
+  const cmnShared = COMMONS.getShared();
+
+  //stats
+  const emote = messageReaction.emoji; //get emote
+  const emoteGuild = emote.guild ? emote.guild : null; //get emote guild if any
+  if (emoteGuild && currentServer.guildId === emoteGuild.id) {
+    //if is a guildEmote and belongs to current server, count
+    const db = messageReaction.client.db;
+    addEmojiData(db, user.id, emote.id); //user stat
+    addServerEmojiCount(db, emote.id); //server stat
+    return;
+  }
+
+  if (
+    currentServer.cosmeticRoleHandle.messageId === messageReaction.message.id
+  ) {
+    roleAdd(messageReaction, currentServer, user);
+    return;
+  }
+
+  if (cmnShared.octagonalSignEmoji === messageReaction.emoji.name) {
+    octagonalLog(messageReaction, user);
+    return;
+  }
+
+  if (
+    messageReaction.message.channel.id ===
+      currentServer.presentationChannelId &&
+    currentServer.presentationReactId === messageReaction.emoji.name
+  ) {
+    console.log("detected");
+    presentationHandler(currentServer, messageReaction, user);
+    return; //no command in presentation channel
+  }
+
+  onRemoveSpotifyReaction(messageReaction, cmnShared);
+
+  onRemoveReminderReaction(messageReaction, user, cmnShared);
+};
+
+export const onReactionRemove = async (messageReaction, user) => {
+  const currentServer = COMMONS.fetchFromGuildId(
+    messageReaction.message.channel.guild.id
+  );
+
+  if (currentServer.cosmeticRoleHandle.messageId === messageReaction.message.id)
+    await roleRemove(messageReaction, currentServer, user);
+};
+
+//#endregion
+
+//#region Listeners helpers
+
+const onPublicMessage = (message, currentServer) => {
   const { author } = message;
 
   if (
@@ -22,7 +158,15 @@ export const onPublicMessage = (message, client, currentServer) => {
   )
     return;
 
-  reactionHandler(message, currentServer, client);
+  if (
+    message.attachments.size &&
+    message.channel.id === currentServer.catsThreadId
+  )
+    addServerStatsData(message.client.db, "cats");
+
+  readContentAndReact(message, currentServer);
+  statsGifCount(message);
+  emojiInContentHandler(message);
 };
 
 export const onRemoveReminderReaction = (
@@ -98,47 +242,4 @@ export const onRemoveSpotifyReaction = async (messageReaction, cmnShared) => {
   }
 };
 
-// Partial ADMIN
-
-export const onReactionAdd = async (messageReaction, user) => {
-  // Function triggered for each reaction added
-  const currentServer = COMMONS.fetchGuildId(
-    messageReaction.message.channel.guild.id
-  );
-  const cmnShared = COMMONS.getShared();
-
-  if (
-    currentServer.cosmeticRoleHandle.messageId === messageReaction.message.id
-  ) {
-    roleAdd(messageReaction, currentServer, user);
-    return;
-  }
-
-  if (cmnShared.octagonalSignEmoji === messageReaction.emoji.name) {
-    octagonalLog(messageReaction, user);
-    return;
-  }
-
-  if (
-    messageReaction.message.channel.id ===
-      currentServer.presentationChannelId &&
-    currentServer.presentationReactId === messageReaction.emoji.name
-  ) {
-    console.log("detected");
-    presentationHandler(currentServer, messageReaction, user);
-    return; //no command in presentation channel
-  }
-
-  onRemoveSpotifyReaction(messageReaction, cmnShared);
-
-  onRemoveReminderReaction(messageReaction, user, cmnShared);
-};
-
-export const onReactionRemove = async (messageReaction, user) => {
-  const currentServer = COMMONS.fetchGuildId(
-    messageReaction.message.channel.guild.id
-  );
-
-  if (currentServer.cosmeticRoleHandle.messageId === messageReaction.message.id)
-    await roleRemove(messageReaction, currentServer, user);
-};
+//#endregion