+++ changelog.md
@@ -1,4 +1,57 @@
-﻿## 6.0.0
+﻿## 6.1.0
+- Fix: `activity` not using `ActivityType` ([#278](https://github.com/Eccleria/ewibot/issues/278)) ([#279](https://github.com/Eccleria/ewibot/pull/279))
+- Fix: `/vest` ([#280](https://github.com/Eccleria/ewibot/pull/280))
+	- missing credits in `/help`
+	- wrong behaviour on vest send
+- Update: reduce `activity` timeout max duration ([#279](https://github.com/Eccleria/ewibot/pull/279))
+
+## 6.0.0
+- Core: reorganize functions and methods for consistency ([#247](https://github.com/Eccleria/ewibot/pull/247))
+	- admin/commands utils that are shared in another folder were moved in helpers utils file
+	- `Commons` class and some functions (`setupEmbed`, `finishEmbed`) were simplified
+	- `listeners` are now more consistent (`onInteractionCreate`, `onMessageCreate` moved)
+	- a lot of function names were modified for easier understanding
+	- `concrete` folder were move in pics, for future pictures fonctionalities
+	- simplify `bot` main file
+	- add jest for future unit testing and auto tests
+	- public `text` are now in personality
+- Doc:
+	- update `readme` with recent and older changes ([#247](https://github.com/Eccleria/ewibot/pull/247)) ([#245](https://github.com/Eccleria/ewibot/pull/245))
+	- add more docstrings to functions/methods ([#247](https://github.com/Eccleria/ewibot/pull/247))
+	- add `canvas` (`concrete` + `vest`) + `helpers` + `announce` doc ([#259](https://github.com/Eccleria/ewibot/pull/259)) ([#245](https://github.com/Eccleria/ewibot/pull/245))
+- Fix: wrong behaviour if only one time value is set for `poll` timeout ([#276](https://github.com/Eccleria/ewibot/issues/276))
+- Feat: add admin `/reaction` command to make the bot react to designated message ([#251](https://github.com/Eccleria/ewibot/pull/251))
+- Feat: add `vest` command, creating a personnalized `vest` with `user` data ([#259](https://github.com/Eccleria/ewibot/pull/259))
+- Feat: add admin `/personality` command ([#270](https://github.com/Eccleria/ewibot/pull/270))
+- Feat: add `poll` ending time in a new embed ([#277](https://github.com/Eccleria/ewibot/issues/277))
+- Update: `concrete` and `vest` store pp hash in pic name for easier pic update ([#260](https://github.com/Eccleria/ewibot/pull/260))
+
+## 5.5.3
+- Fix: wrong uptade value for PR ([#269](https://github.com/Eccleria/ewibot/pull/269)) ([#271](https://github.com/Eccleria/ewibot/pull/271)). ([#274](https://github.com/Eccleria/ewibot/pull/274))
+- Update: remove `messageUpdate` length filter which remove some important logs ([#269](https://github.com/Eccleria/ewibot/pull/269))
+- Update: `gift` ([#268](https://github.com/Eccleria/ewibot/issues/268)) ([#271](https://github.com/Eccleria/ewibot/pull/271))
+	- disable older buttons using a yearly `customId` recognition.
+	- update dates to 2023 case
+	- `/gift accepting` now display full tag list if no `recipient` argument
+
+## 5.5.2
+- Feat: optional `activity` for Halloween ([#262](https://github.com/Eccleria/ewibot/issues/262)) ([#263](https://github.com/Eccleria/ewibot/pull/263)) ([#264](https://github.com/Eccleria/ewibot/pull/264))
+
+## 5.5.1
+- Fix: wrong `author` on some `onThreadCreate` log when `thread` is created ([#252](https://github.com/Eccleria/ewibot/issues/252)) ([#253](https://github.com/Eccleria/ewibot/pull/253))
+- Fix: multiple `poll` issues ([#240](https://github.com/Eccleria/ewibot/issues/240)) ([#254](https://github.com/Eccleria/ewibot/pull/254))
+	- error if user vote during `poll` closure
+	- difficulties on `emote` detection in `poll` choice submission
+	- `dbPoll` `undefined` if accessed after `poll` closure
+- Update: `/roll` result is now public ([#261](https://github.com/Eccleria/ewibot/pull/261))
+
+## 5.5.0
+- Feat: statistics ([#234](https://github.com/Eccleria/ewibot/pull/234))
+	- add `/stats use` command to dis/enable Ewibot stats for requesting user
+	- count `gifs`, `hello` words, `hungry` words, and `emojis` for accepting users
+	- count `emotes` occurences and `cats` pictures sent for whole guild
+- Feat: `dbReturnType` frozen object with returned values for db helpers ([#234](https://github.com/Eccleria/ewibot/pull/234))
+- Fix: alavirien wrong db access when alavirien check ([#250](https://github.com/Eccleria/ewibot/pull/234))
 
 ## 5.4.0
 - Feat: add reaction if is mentioned by a user ([#230](https://github.com/Eccleria/ewibot/pull/230))